refactor: value 생성자 주입으로 변경
