+++ wnp-server/src/main/java/com/wonnapark/wnpserver/media/S3ManageService.java
@@ -1,8 +1,7 @@
-package com.wonnapark.wnpserver.domain.infra;
+package com.wonnapark.wnpserver.media;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.wonnapark.wnpserver.global.utils.FileUtils;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -11,13 +10,19 @@ import java.io.File;
 import java.io.IOException;
 
 @Service
-@RequiredArgsConstructor
 public class S3ManageService {
 
     private final AmazonS3 amazonS3;
 
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucketName;
+    private final String bucketName;
+
+    public S3ManageService(
+            AmazonS3 amazonS3,
+            @Value("${cloud.aws.s3.bucket}") String bucketName
+    ) {
+        this.amazonS3 = amazonS3;
+        this.bucketName = bucketName;
+    }
 
     public void upload(String key, MultipartFile file) throws IOException {
         amazonS3.putObject(bucketName, key, FileUtils.convertMultipartFileToFile(file));
